// core/admin-dashboard/store/feedback/action-types.ts
export enum ActionTypes {
  GET_FEEDBACKS = "GET_FEEDBACKS",
  GET_FEEDBACK = "GET_FEEDBACK",
  CREATE_FEEDBACK = "CREATE_FEEDBACK",
  UPDATE_FEEDBACK = "UPDATE_FEEDBACK",
  DELETE_FEEDBACK = "DELETE_FEEDBACK",
  UPLOAD_FEEDBACK_THUMBNAIL = "UPLOAD_FEEDBACK_THUMBNAIL",
  HARD_DELETE_FEEDBACK = "HARD_DELETE_FEEDBACK",
  GET_FEEDBACK_ANALYTICS = "GET_FEEDBACK_ANALYTICS",
}